refactor: delegate leaderboard sorting to repository for better performance
