+++ pom.xml
@@ -10,7 +10,7 @@
     </parent>
     <groupId>org.example</groupId>
     <artifactId>WordleApp</artifactId>
-    <version>1.5.2</version>
+    <version>1.5.4</version>
     <name>WordleApp</name>
     <description>WordleApp</description>
     <properties>
