+++ src/main/java/com/wordleapp/repository/GameRepository.java
@@ -7,5 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface GameRepository extends JpaRepository<Game, Long> {
+    List<Game> findAllByOrderByAttemptsAsc();
+    List<Game> findAllByOrderByCreatedAtDesc();
     List<Game> findBySecretWordOrderByAttempts(SecretWord secretWord);
+
 }
