+++ src/main/java/com/wordleapp/service/LeaderboardService.java
@@ -18,14 +18,15 @@ public class LeaderboardService {
     }
 
     public List<LeaderboardEntry> getLeaderboard(String orderBy) {
-        Comparator<Game> comparator = Comparator.comparing(Game::getCreatedAt).reversed();
+        List<Game> games;
 
         if ("attempts".equalsIgnoreCase(orderBy)) {
-            comparator = Comparator.comparingInt(Game::getAttempts);
+            games = gameRepository.findAllByOrderByAttemptsAsc();
+        } else {
+            games = gameRepository.findAllByOrderByCreatedAtDesc();
         }
 
-        return gameRepository.findAll().stream()
-                .sorted(comparator)
+        return games.stream()
                 .map(game -> new LeaderboardEntry(
                         game.getUsername(),
                         game.getSecretWord().getWord(),
@@ -34,4 +35,5 @@ public class LeaderboardService {
                 ))
                 .toList();
     }
+
 }
