+++ src/test/java/com/wordleapp/integration/PageControllerTest.java
@@ -1,4 +1,4 @@
-package com.wordleapp.unit;
+package com.wordleapp.integration;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
